refactor: id 타입 long 변경

+++ src/main/java/org/devcourse/voucher/controller/console/dto/VoucherInfoResponse.java
@@ -1,4 +1,4 @@
 package org.devcourse.voucher.controller.console.dto;
 
-public record VoucherInfoResponse(int id, String voucherType, int amount) {
+public record VoucherInfoResponse(long id, String voucherType, int amount) {
 }
